requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
django-filter==24.3
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
django-allauth>=65.11.2
djangorestframework-simplejwt==5.3.0
setuptools>=80.9.0
requests==2.32.5
PyJWT==2.10.1
django-cors-headers==4.9.0
python-decouple==3.8
cryptography==46.0.1
django-cors-headers==4.9.0
